+++ RELEASE_NOTES.md
@@ -13,6 +13,10 @@
 1.  Remove unnecessary dependency of map assets on Lua scripts, preventing
     time-consuming rebuilding of maps when scripts are modified.
 2.  Add ability to disable bobbing of reward and goal pickups.
+3.  The setting `controls` (with values `internal`, `external`) has been renamed
+    to `nativeApp` (with values `true`, `false`, respectively). When set to
+    `true`, programs linked against `game_lib_sdl` will use the native SDL input
+    devices.
 
 ### Bug Fixes: